requirements.txt
python-chess==1.999
requests==2.31.0
colorama==0.4.6
